requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
coverage==7.2.7
dj-config-url==0.1.1
dj-database-url==2.0.0
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
Markdown==3.4.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
Pygments==2.15.1
pytest==7.4.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
